BatchNorm and FixedBatchNorm op registration
