+++ chainerx_cc/chainerx/cuda/cuda_device.h
@@ -219,15 +219,6 @@ public:
             const StackVector<int64_t, kMaxNdim>& pad,
             AveragePoolPadMode pad_mode) override;
 
-    // batch_norm.cc
-
-    std::unique_ptr<BatchNormForwardBackward> GetBatchNormForwardBackward(
-            const Array& running_mean, const Array& running_var, Scalar eps, Scalar decay, const Axes& axis) override;
-
-    Array FixedBatchNorm(
-            const Array& x, const Array& gamma, const Array& beta, const Array& mean, const Array& var, Scalar eps, const Axes& axis)
-            override;
-
 protected:
     CudaDevice(CudaBackend& backend, int index)
         : Device{backend, index},
