+++ chainerx_cc/chainerx/native/CMakeLists.txt
@@ -15,6 +15,7 @@ add_library(chainerx_native STATIC
     native_device.cc
     native_device/activation.cc
     native_device/arithmetic.cc
+    native_device/batch_norm.cc
     native_device/conv.cc
     native_device/copy.cc
     native_device/comparison.cc
